requirements.txt
openai
fpdf
python-dotenv
requests
beautifulsoup4
